refine package structure under 'org.mozilla.rocket.vertical'
